[fix] correct update_screen selected entry render behaviour
When the user selects some files in visual mode, then the marking of
those files that are  within the first pane length (0 to term.rows) are
always rendered at their first pane position when the user calls
update_screen. If the user selects file entries that are beyond the
first pane length, then update_screen never or unrenders them.

This modification eliminates this behaviour.